urban_infra_project/urban_infra/requirements.txt
Django>=4.2,<5.0
Pillow>=10.0.0
django-environ>=0.11.2
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
django-allauth>=0.58.2
djangorestframework>=3.14.0
django-filter>=23.3
openpyxl>=3.1.2
reportlab>=4.0.7
python-docx>=1.1.0
celery>=5.3.4
redis>=5.0.1
channels>=4.0.0
channels-redis>=4.1.0
django-cors-headers>=4.3.1
django-extensions>=3.2.3
whitenoise>=6.6.0
python-decouple>=3.8
requests>=2.31.0
pyotp>=2.9.0
qrcode>=7.4.2
django-ckeditor>=6.7.0
